auth-service/requirements.txt
django
djangorestframework
PyJWT
PyMySQL


cryptography
